@@ -4,4 +4,8 @@
 requires = ["setuptools", "wheel"]
 
 [tool.black]
-target-version = ['py35', 'py36', 'py37', 'py38']
+target-version = ['py36']
+
+[tool.isort]
+profile = "black"
+src_paths = ["pytest_plt"]
