@@ -3,7 +3,7 @@
 # Automatically generated by nengo-bones, do not edit this file directly
 
 import io
-import os
+import pathlib
 import runpy
 
 try:
@@ -26,8 +26,8 @@ def read(*filenames, **kwargs):
     return sep.join(buf)
 
 
-root = os.path.dirname(os.path.realpath(__file__))
-version = runpy.run_path(os.path.join(root, "pytest_plt", "version.py"))["version"]
+root = pathlib.Path(__file__).parent
+version = runpy.run_path(str(root / "pytest_plt" / "version.py"))["version"]
 
 install_req = [
     "matplotlib",
@@ -61,15 +61,20 @@ setup(
         "optional": optional_req,
         "tests": tests_req,
     },
-    python_requires=">=3.5",
-    entry_points={"pytest11": ["plt = pytest_plt.plugin",],},
+    python_requires=">=3.6",
+    entry_points={
+        "pytest11": [
+            "plt = pytest_plt.plugin",
+        ],
+    },
     classifiers=[
         "Development Status :: 5 - Production/Stable",
         "Framework :: Pytest",
         "License :: OSI Approved :: MIT License",
         "Programming Language :: Python :: 3 :: Only",
-        "Programming Language :: Python :: 3.5",
         "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
+        "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
     ],
 )
