@@ -2,7 +2,7 @@
 #
 # Automatically generated by nengo-bones, do not edit this file directly
 
-import os
+import pathlib
 
 import pytest_plt
 
@@ -14,10 +14,11 @@ extensions = [
     "sphinx.ext.intersphinx",
     "sphinx.ext.mathjax",
     "sphinx.ext.todo",
-    "sphinx.ext.viewcode",
     "nbsphinx",
     "nengo_sphinx_theme",
     "nengo_sphinx_theme.ext.backoff",
+    "nengo_sphinx_theme.ext.sourcelinks",
+    "notfound.extension",
     "numpydoc",
 ]
 
@@ -34,18 +35,26 @@ import pytest_plt
 # -- sphinx.ext.intersphinx
 intersphinx_mapping = {
     "nengo": ("https://www.nengo.ai/nengo/", None),
-    "numpy": ("https://docs.scipy.org/doc/numpy", None),
+    "numpy": ("https://numpy.org/doc/stable", None),
     "python": ("https://docs.python.org/3", None),
 }
 
 # -- sphinx.ext.todo
 todo_include_todos = True
 
+# -- nbsphinx
+nbsphinx_timeout = -1
+
+# -- notfound.extension
+notfound_template = "404.html"
+notfound_urls_prefix = "/pytest-plt/"
+
 # -- numpydoc config
 numpydoc_show_class_members = False
 
-# -- nbsphinx
-nbsphinx_timeout = -1
+# -- nengo_sphinx_theme.ext.sourcelinks
+sourcelinks_module = "pytest_plt"
+sourcelinks_url = "https://github.com/nengo/pytest-plt"
 
 # -- sphinx
 nitpicky = True
@@ -65,7 +74,7 @@ user_agent = "pytest_plt"
 
 project = "pytest-plt"
 authors = "Applied Brain Research"
-copyright = "2018-2020 Applied Brain Research"
+copyright = "2018-2021 Applied Brain Research"
 version = ".".join(pytest_plt.__version__.split(".")[:2])  # Short X.Y version
 release = pytest_plt.__version__  # Full version, with tags
 
@@ -73,14 +82,14 @@ release = pytest_plt.__version__  # Full version, with tags
 templates_path = ["_templates"]
 html_static_path = ["_static"]
 html_theme = "nengo_sphinx_theme"
-html_title = "pytest-plt {0} docs".format(release)
+html_title = f"pytest-plt {release} docs"
 htmlhelp_basename = "pytest-plt"
 html_last_updated_fmt = ""  # Default output format (suppressed)
 html_show_sphinx = False
-html_favicon = os.path.join("_static", "favicon.ico")
+html_favicon = str(pathlib.Path("_static", "favicon.ico"))
 html_theme_options = {
     "nengo_logo": "general-square-light.svg",
     "nengo_logo_color": "#a8acaf",
-    "analytics_id": "UA-41658423-2",
+    "tagmanager_id": "GTM-KWCR2HN",
     "one_page": True,
 }
