@@ -1,16 +1,24 @@
+# Automatically generated by nengo-bones, do not edit this file directly
+
+# pylint: disable=consider-using-f-string
+
 """pytest-plt version information.
 
-We use semantic versioning (see http://semver.org/).
+We use semantic versioning (see http://semver.org/)
 and conform to PEP440 (see https://www.python.org/dev/peps/pep-0440/).
-'.devN' will be added to the version unless the code base represents
+'.dev0' will be added to the version unless the code base represents
 a release version. Release versions are git tagged with the version.
 """
 
-name = "pytest_plt"
-version_info = (1, 1, 1)  # (major, minor, patch)
+version_info = (1, 1, 1)
+
+name = "pytest-plt"
 dev = 0
 
-version = "{v}{dev}".format(
-    v=".".join(str(v) for v in version_info),
-    dev=(".dev%d" % dev) if dev is not None else "",
-)
+# use old string formatting, so that this can still run in Python <= 3.5
+# (since this file is parsed in setup.py, before python_requires is applied)
+version = ".".join(str(v) for v in version_info)
+if dev is not None:
+    version += ".dev%d" % dev
+
+copyright = "Copyright (c) 2018-2021 Applied Brain Research"
